replace float32 with float64 for OpsLevel terraform provider (#395)
* replace float32 with float64 for OpsLevel terraform provider

* Update .changes/unreleased/Refactor-20240502-084345.yaml


---------

+++ repository.go
@@ -9,7 +9,7 @@ import (
 
 type Language struct {
 	Name  string
-	Usage float32
+	Usage float64
 }
 
 type RepositoryId struct {
